{{cookiecutter.project_slug}}/requirements.txt
pulsar-client==2.5.1
tanbih-pipeline==0.1.*
pytest
